requirements.txt
fastapi==0.103.1
azure-functions==1.17.0
uvicorn==0.23.2
python-multipart==0.0.6